Updated order service files

+++ Order-service/src/main/java/com/example/Orderservice/Dto/OrderDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class OrderDTO {
     private Long id;
     private Integer status;
-    private Long subtotal;
+    private Double subtotal;
     private Date createdAt;
     private Date updatedAt;
     private Long userId;
